+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 - Changes `superstatic` dependency to `v8`, addressing Hosting emulator issues on Windows.
 - Fixes internal library that was not being correctly published.
 - Adds `--disable-triggers` flag to RTDB write commands.
-- Fixes source token expiration issue by refreshing source token.
+- Fixes source token expiration issue by acquiring new source token upon expiration.